REFACTOR CMSPageAddControllerExtension updateDoAdd()
Previously only worked if the ElementalArea relation was named `ElementalArea`.

This will find the first ElementalArea for the class that has a field in getCMSFields(). For example, on `Dynamic\Base\Page\HomePage`, we add has_one `ElementalHomePage` and hide the `ElementalArea` field. In this case, blocks would be copied to `ElementalHomePage`