REFACTOR Improve support for custom ElementalArea relations in page templates.
Previously, creating a page from a template only worked if the ElementalArea relation was named ElementalArea. This update finds the first ElementalArea relation for the class that has a corresponding field in getCMSFields().

For example, on Dynamic\Base\Page\HomePage, where ElementalHomePage is defined as a has_one and ElementalArea is hidden, blocks will now be copied to ElementalHomePage.